Don't use genesis node to create VMs

+++ vars/heat.groovy
@@ -1,4 +1,3 @@
-
 // Required credentials names
 //  - jenkins-openstack
 //  - jenkins-token
@@ -10,17 +9,23 @@
 //  - JENKINS_CLI
 
 
-def openstack_cmd(String cmd, String mount = "") {
+def openstack_cmd(String cmd, String mount = "", Boolean useHeatContainer = true) {
+    openstack_credentials = [
+        'OS_AUTH_URL': OS_AUTH_URL,
+        'OS_PROJECT_DOMAIN_NAME': 'default',
+        'OS_USER_DOMAIN_NAME': 'default',
+        'OS_PROJECT_NAME': OS_PROJECT_NAME,
+        'OS_REGION_NAME': OS_REGION_NAME,
+        'OS_USERNAME': OS_USERNAME,
+        'OS_PASSWORD': OS_PASSWORD,
+        'OS_IDENTITY_API_VERSION':'3'
+    ]
 
-    docker_env = " -e OS_AUTH_URL=${OS_AUTH_URL}" +
-                 " -e OS_PROJECT_DOMAIN_NAME=default" +
-                 " -e OS_USER_DOMAIN_NAME=default" +
-                 " -e OS_PROJECT_NAME=${OS_PROJECT_NAME}" +
-                 " -e OS_REGION_NAME=${OS_REGION_NAME}" +
-                 " -e OS_USERNAME=\$OS_USERNAME" +
-                 " -e OS_PASSWORD=\$OS_PASSWORD" +
-                 " -e OS_IDENTITY_API_VERSION=3"
+    if (useHeatContainer == false) {
+        return "${openstack_credentials.collect { "${it.key}=${it.value}" }.join(' ')} $cmd"
+    }
 
+    docker_env = openstack_credentials.collect { "-e ${it.key}=${it.value}" }.join(' ')
     docker_opts = "--rm --net=host"
 
     if (mount) {
@@ -31,31 +36,35 @@ def openstack_cmd(String cmd, String mount = "") {
 }
 
 //example of parameters --parameter 'cloudImage=${cloudImage}' --parameter ...
-def stack_create(String name, String tmpl, String parameters) {
+def stack_create(String name, String tmpl, String parameters, Boolean useHeatContainer = true) {
 
   withCredentials([usernamePassword(credentialsId: 'jenkins-openstack-18',
                                           usernameVariable: 'OS_USERNAME',
                                           passwordVariable: 'OS_PASSWORD')]) {
-
-        cmd = openstack_cmd("openstack stack create -t /target/\$(basename ${tmpl}) ${name} ${parameters}", "\$(dirname ${tmpl})")
+        String cmd
+        if (useHeatContainer) {
+            cmd = openstack_cmd("openstack stack create -t '/target/\$(basename ${tmpl})' '${name}' ${parameters} --wait", "\$(dirname ${tmpl})")
+        } else {
+            cmd = openstack_cmd("openstack stack create -t '${tmpl}' '${name}' ${parameters} --wait", null, false)
+        }
         code = sh (script: cmd, returnStatus: true)
         if (!code) {
             // todo: improve timeouts to more user friendly look
             timeout = 300
             for (i = 0; i < timeout; i=i+10) {
-                sleep 30
-                cmd = openstack_cmd("openstack stack show -f value -c stack_status ${name}")
+                cmd = openstack_cmd("openstack stack show -f value -c stack_status ${name}", null, useHeatContainer)
                 ret = sh (script: cmd, returnStdout: true).trim()
                 if (ret == "CREATE_COMPLETE") {
                     print "Stack ${name} created!"
                     return
                 } else if (ret != "CREATE_IN_PROGRESS") {
-                    cmd = openstack_cmd("openstack stack show ${name}")
+                    cmd = openstack_cmd("openstack stack show ${name}", null, useHeatContainer)
                     ret = sh (script: cmd, returnStdout: true)
                     print "Stack status:\n${ret}"
                     print "Heat stack error ${name}"
                     sh "exit 1"
                 }
+                sleep 30
             }
         }
         print "Failed to create stack ${name}"
@@ -64,15 +73,14 @@ def stack_create(String name, String tmpl, String parameters) {
 }
 
 
-def stack_delete(String name) {
+def stack_delete(String name, Boolean useHeatContainer = true) {
     withCredentials([usernamePassword(credentialsId: 'jenkins-openstack-18',
                                           usernameVariable: 'OS_USERNAME',
                                           passwordVariable: 'OS_PASSWORD')]) {
-
-      cmd = openstack_cmd("openstack stack delete --wait --yes ${name}")
+        cmd = openstack_cmd("openstack stack delete --wait --yes ${name}", null, useHeatContainer)
         code = sh (script: cmd, returnStatus: true)
         if (!code) {
-            cmd = openstack_cmd("openstack stack list")
+            cmd = openstack_cmd("openstack stack list", null, useHeatContainer)
             ret = sh (script: cmd, returnStdout: true)
             if (!ret.contains(name)) {
                 print "Stack ${name} deleted!"
@@ -87,17 +95,17 @@ def stack_delete(String name) {
 }
 
 
-def stack_output(String name, String output) {
+def stack_output(String name, String output, Boolean useHeatContainer = true) {
     withCredentials([usernamePassword(credentialsId: 'jenkins-openstack-18',
                                       usernameVariable: 'OS_USERNAME',
                                       passwordVariable: 'OS_PASSWORD')]) {
-        cmd = openstack_cmd("openstack stack output show -f value -c output_value ${name} ${output}")
+        cmd = openstack_cmd("openstack stack output show -f value -c output_value ${name} ${output}", null, useHeatContainer)
         return sh(returnStdout: true, script: cmd).trim()
     }
 }
 
-def stack_status(String name) {
-    cmd = openstack_cmd("openstack stack show -f value -c stack_status ${name}")
+def stack_status(String name, Boolean useHeatContainer = true) {
+    cmd = openstack_cmd("openstack stack show -f value -c stack_status ${name}", null, useHeatContainer)
     ret = sh (script: cmd, returnStdout: true).trim()
     if (ret != "CREATE_COMPLETE") {
         print "Failed to create stack ${name}"
